Restructure projects app models
